# specs/001-p2p-payment-request/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: P2P Payment Request API
  version: 0.1.0
  description: |
    Prototype API for peer-to-peer payment requests.
    Authentication is mock email-based via X-User-Email header after login.
    Production auth is out of scope.
servers:
  - url: http://localhost:8000
    description: Local FastAPI (dev or single-service)
...
components:
  securitySchemes:
    UserEmailHeader:
      type: apiKey
      in: header
      name: X-User-Email
      description: Prototype auth — set to logged-in user email
  schemas:
    ErrorResponse:
      type: object
      required: [detail]
      properties:
        detail:
          oneOf:
            - type: string
            - type: array
              items:
                type: object
    User:
      type: object
      required: [id, email, created_at]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        display_name: { type: string, nullable: true }
        phone: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 1, maxLength: 128, example: "1234" }
    LoginResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: '#/components/schemas/User'
...
paths:
  /api/auth/login:
    post:
      tags: [auth]
      summary: Mock email login or register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: User logged in; default wallet and destination ensured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '422':
          description: Invalid email
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/me:
    get:
      tags: [auth]
      security: [{ UserEmailHeader: [] }]
      summary: Current user profile
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Missing or unknown user header
...